BitsAndByte-Backend/requirements.txt
asgiref==3.2.10
certifi==2019.11.28
chardet==3.0.4
Django==3.1.5
gunicorn==20.0.4
idna==2.8
psycopg2==2.8.6
pytz==2019.3
requests==2.22.0
sqlparse==0.2.4
urllib3==1.25.8
whitenoise==5.2.0
django-heroku==0.3.1
django-cors-headers==3.7.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
PyJWT==2.1.0
qrcode==7.0
requests==2.22.0
Pillow==5.2.0
